[A] 1. 新增个人中心 2. 头像上传暂时有BUG

+++ src/app/pages/main/user-management/user-list/user-list.component.ts
@@ -52,6 +52,10 @@ export class UserListComponent implements OnInit {
 
   // 删除用户
   deleteUser(id: number): void {
+    if (id === SessionUtil.getUserId()) {
+      this.$message.error('无法删除自己');
+      return;
+    }
     this.userManagementRequestService.deleteUser(id).subscribe((result: Result<void>) => {
       if (result.code === 200) {
         this.$message.success(result.msg);
